Added agency service Authservice and DuplicateResourceNotFoundExcepti…

+++ src/main/java/com/busbooking/repository/AgencyOfficeRepo.java
@@ -4,10 +4,9 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.busbooking.entity.Agency;
 import com.busbooking.entity.AgencyOffice;
 
-public interface AgencyOfficeRepo extends JpaRepository<AgencyOffice, Integer> {
+public interface AgencyOfficeRepo extends JpaRepository<AgencyOffice, Long> {
 	
 	List<AgencyOffice> findByAgency_AgencyId(Long agencyId);
